- changed timestamp type from BigInt to Int - removed unnecesary params on signer function - changed DEFAULT_DECIMALS to prevent error on deploying on gnosis chain when trying to convert it toI32()

+++ src/modules/totals.ts
@@ -22,7 +22,7 @@ export namespace totals {
         return total as Total
     }
 
-    function getOrCreateDailyTotals(timestamp: BigInt): DailyTotal {
+    function getOrCreateDailyTotals(timestamp: i32): DailyTotal {
         let totalId = timestamp.toString()
         let total = DailyTotal.load(totalId)
 
@@ -42,7 +42,7 @@ export namespace totals {
         return total as DailyTotal
     }
 
-    function getOrCreateHourlyTotals(timestamp: BigInt): HourlyTotal {
+    function getOrCreateHourlyTotals(timestamp: i32): HourlyTotal {
         let totalId = timestamp.toString()
         let total = HourlyTotal.load(totalId)
 
@@ -70,7 +70,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateDailyTokensCount(timestamp: BigInt, tokenId: string): void {
+    function updateDailyTokensCount(timestamp: i32, tokenId: string): void {
         let total = getOrCreateDailyTotals(timestamp)
         let prevtotalTokens = total.totalTokens
         let prevTokens = total.tokens
@@ -80,7 +80,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateHourlyTokensCount(timestamp: BigInt, tokenId: string): void {
+    function updateHourlyTokensCount(timestamp: i32, tokenId: string): void {
         let total = getOrCreateHourlyTotals(timestamp)
         let prevtotalTokens = total.totalTokens
         let prevTokens = total.tokens
@@ -90,7 +90,7 @@ export namespace totals {
         total.save()
     }
 
-    export function addTokenCount(timestamp: BigInt, tokenId: string): void{
+    export function addTokenCount(timestamp: i32, tokenId: string): void{
         let dayTimestamp = getDayTotalTimestamp(timestamp)
         let hourTimestamp = getHourTotalTimestamp(timestamp)
         updateDailyTokensCount(dayTimestamp, tokenId)
@@ -105,7 +105,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateHourlyOrdersCount(timestamp: BigInt): void {
+    function updateHourlyOrdersCount(timestamp: i32): void {
         let total = getOrCreateHourlyTotals(timestamp)
         let prevtotalOrders = total.orders
 
@@ -113,7 +113,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateDailyOrdersCount(timestamp: BigInt): void {
+    function updateDailyOrdersCount(timestamp: i32): void {
         let total = getOrCreateDailyTotals(timestamp)
         let prevtotalOrders = total.orders
 
@@ -121,7 +121,7 @@ export namespace totals {
         total.save()
     }
 
-    export function addOrderCount(timestamp: BigInt): void {
+    export function addOrderCount(timestamp: i32): void {
         let dayTimestamp = getDayTotalTimestamp(timestamp)
         let hourTimestamp = getHourTotalTimestamp(timestamp)
         updateDailyOrdersCount(dayTimestamp)
@@ -143,7 +143,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateDailySettlementsCount(timestamp: BigInt): void {
+    function updateDailySettlementsCount(timestamp: i32): void {
         let total = getOrCreateDailyTotals(timestamp)
         let prevtotalSettlements = total.settlements
 
@@ -151,7 +151,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateHourlySettlementsCount(timestamp: BigInt): void {
+    function updateHourlySettlementsCount(timestamp: i32): void {
         let total = getOrCreateHourlyTotals(timestamp)
         let prevtotalSettlements = total.settlements
 
@@ -159,7 +159,7 @@ export namespace totals {
         total.save()
     }
 
-    export function addSettlementCount(timestamp: BigInt): void {
+    export function addSettlementCount(timestamp: i32): void {
         let dayTimestamp = getDayTotalTimestamp(timestamp)
         let hourTimestamp = getHourTotalTimestamp(timestamp)
         updateDailySettlementsCount(dayTimestamp)
@@ -181,7 +181,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateHourlyVolumesAndFees(volumeEth: BigDecimal, volumeUsd: BigDecimal, feesEth: BigDecimal, feesUsd: BigDecimal, timestamp: BigInt): void {
+    function updateHourlyVolumesAndFees(volumeEth: BigDecimal, volumeUsd: BigDecimal, feesEth: BigDecimal, feesUsd: BigDecimal, timestamp: i32): void {
         let total = getOrCreateHourlyTotals(timestamp)
         let prevVolumeEth = total.volumeEth
         let prevVolumeUsd = total.volumeUsd
@@ -194,7 +194,7 @@ export namespace totals {
         total.save()
     }
 
-    function updateDailyVolumesAndFees(volumeEth: BigDecimal, volumeUsd: BigDecimal, feesEth: BigDecimal, feesUsd: BigDecimal, timestamp: BigInt): void {
+    function updateDailyVolumesAndFees(volumeEth: BigDecimal, volumeUsd: BigDecimal, feesEth: BigDecimal, feesUsd: BigDecimal, timestamp: i32): void {
         let total = getOrCreateDailyTotals(timestamp)
         let prevVolumeEth = total.volumeEth
         let prevVolumeUsd = total.volumeUsd
@@ -207,7 +207,7 @@ export namespace totals {
         total.save()
     }
 
-    export function addVolumesAndFees(volumeEth: BigDecimal, volumeUsd: BigDecimal, feesEth: BigDecimal, feesUsd: BigDecimal, timestamp: BigInt): void {
+    export function addVolumesAndFees(volumeEth: BigDecimal, volumeUsd: BigDecimal, feesEth: BigDecimal, feesUsd: BigDecimal, timestamp: i32): void {
         let dayTimestamp = getDayTotalTimestamp(timestamp)
         let hourTimestamp = getHourTotalTimestamp(timestamp)
         updateDailyVolumesAndFees(volumeEth, volumeUsd, feesEth, feesUsd, dayTimestamp)
